Rename `compat-target28` integration test to `sdkcompat`
`sdkcompat` is a more generic and relevant name.
