+++ integration_tests/sdkcompat/build.gradle
@@ -15,7 +15,7 @@ spotless {
 
 android {
     compileSdk 28
-    namespace 'org.robolectric.integrationtests.compattarget28'
+    namespace 'org.robolectric.integrationtests.sdkcompat'
 
     defaultConfig {
         minSdk 19
