+++ integration_tests/sdkcompat/src/main/AndroidManifest.xml
@@ -2,6 +2,6 @@
 <manifest xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android">
     <application
-        android:appComponentFactory="org.robolectric.integrationtests.compattarget28.TestAppComponentFactory"
+        android:appComponentFactory="org.robolectric.integrationtests.sdkcompat.TestAppComponentFactory"
         tools:targetApi="p" />
 </manifest>
