+++ integration_tests/sdkcompat/src/test/java/org/robolectric/integrationtests/sdkcompat/JavaClassResolveCompatibilityTest.java
@@ -1,4 +1,4 @@
-package org.robolectric.integration.compat.target28;
+package org.robolectric.integrationtests.sdkcompat;
 
 import static com.google.common.truth.Truth.assertThat;
