+++ integration_tests/sdkcompat/src/test/resources/robolectric.properties
