+++ settings.gradle
@@ -36,7 +36,7 @@ include ':integration_tests:ctesque'
 include ':integration_tests:security-providers'
 include ":integration_tests:mockk"
 include ":integration_tests:jacoco-offline"
-include ':integration_tests:compat-target28'
+include ':integration_tests:sdkcompat'
 include ":integration_tests:multidex"
 include ":integration_tests:play_services"
 include ":integration_tests:sparsearray"
